Dev/search&compare engine

+++ comp_deleter.cpp
@@ -7,7 +7,8 @@ CCompDeleter::CCompDeleter(QWidget *parent) :
   m_pUi(new Ui::CompDeleterWindow)
 {
   m_pUi->setupUi(this);
-  m_hQuery = new CQueryController(CODBCW::getInstance());
+  size_t i = 0;
+  m_hQuery = new CQueryController(CQueryController(CODBCW::getInstance("", nullptr, &i)));
   QWidget::connect(this, &CCompDeleter::finished,
           this, &CCompDeleter::CCompDeleter::clearUi);
 }
@@ -27,33 +28,42 @@ void CCompDeleter::showWindow(QTableView* hTbPriceList) {
 }
 
 void CCompDeleter::on_accept_deleter_btn_clicked() {
-  QString sQuery = QString("DELETE FROM components "
-                           "WHERE title = '%1'")
-                           .arg(m_pUi->comp_name_cmbBox->currentText());
+  QMessageBox::StandardButton reply;
+  reply = QMessageBox::warning(this, "Warning!", "This action may delete order, are you sure? "
+                                                 "More information is available in the help.",
+                               QMessageBox::Yes | QMessageBox::No);
 
-  if (m_hQuery->executeSqlQuery(sQuery)) {
-    CCompsTraits::get_comps().rm(m_pUi->comp_name_cmbBox->currentText());
+  if (reply == QMessageBox::Yes) {
+    QString sQuery = QString("DELETE FROM components "
+                             "WHERE title = '%1'")
+                             .arg(m_pUi->comp_name_cmbBox->currentText());
 
-    QMessageBox::information(this, "Success",
-                             "Component '" + m_pUi->comp_name_cmbBox->currentText() + "' was successfully removed!");
+    if (m_hQuery->executeSqlQuery(sQuery)) {
+      CCompsTraits::get_comps().rm(m_pUi->comp_name_cmbBox->currentText());
 
-    sQuery = "SELECT title AS 'Title',"
-             "specifications AS 'Specifications', price AS 'Price', release_date AS 'Release date' "
-             "FROM components "
-             "JOIN components_type "
-             "ON FK_type_code = PK_component_type_id;";
+      CMsgBox::show(QMessageBox::Information, this, "Success",
+                    "Component '" + m_pUi->comp_name_cmbBox->currentText() + "' was successfully removed!");
 
-    m_hQuery->executeSqlQuery(sQuery);
+      sQuery = "SELECT title AS 'Title', component_type AS 'Component type',"
+               " specifications AS 'Specifications', price AS 'Price', release_date AS 'Release date' "
+               "FROM components "
+               "JOIN components_type "
+               "ON FK_type_code = PK_component_type_id;";
 
-    QSqlQueryModel* hModel = new QSqlQueryModel();
-    QSortFilterProxyModel* hFilterModel = new QSortFilterProxyModel();
-    hModel->setQuery(m_hQuery->getQuery());
-    hFilterModel->setDynamicSortFilter(true);
-    hFilterModel->setSourceModel(hModel);
-    m_hTbPriceList->setModel(hFilterModel);
-    CCompsTraits::get_comps().upd(*m_pUi->comp_name_cmbBox);
-  } else
-    QMessageBox::critical(this, "Error!", "Please check your connection to database!");
+      m_hQuery->executeSqlQuery(sQuery);
+
+      QSqlQueryModel* hModel = new QSqlQueryModel();
+      QSortFilterProxyModel* hFilterModel = new QSortFilterProxyModel();
+      hModel->setQuery(m_hQuery->getQuery());
+      hFilterModel->setDynamicSortFilter(true);
+      hFilterModel->setSourceModel(hModel);
+      m_hTbPriceList->setModel(hFilterModel);
+      CCompsTraits::get_comps().upd(*m_pUi->comp_name_cmbBox);
+
+      emit updOrders();
+    } else
+      CMsgBox::show(QMessageBox::Critical, this, "Error!", "Please check your connection to database!");
+  }
 }
 
 void CCompDeleter::clearUi() {
