+++ comp_type_deleter.cpp
@@ -8,7 +8,9 @@ CComp_TyDeleter::CComp_TyDeleter(QWidget* parent) :
 {
   m_pUi->setupUi(this);
   m_pTbCompType = new QTableView();
-  m_hQuery = new CQueryController(CODBCW::getInstance());
+  m_pTbComps = new QTableView();
+  size_t i = 0;
+  m_hQuery = new CQueryController(CQueryController(CODBCW::getInstance("", nullptr, &i)));
 
   connect(this, &CComp_TyDeleter::finished,
           this, &CComp_TyDeleter::clearUi);
@@ -18,46 +20,76 @@ CComp_TyDeleter::~CComp_TyDeleter() {
   delete m_pUi;
 }
 
-void CComp_TyDeleter::showWindow(QTableView* hTbCompType) {
+void CComp_TyDeleter::showWindow(QTableView* hTbCompType, QTableView* hTbComps) {
   m_pTbCompType = hTbCompType;
+  m_pTbComps = hTbComps;
   CCompsTraits::get_comps().upd(*m_pUi->comp_type_cBox);
   show();
 }
 
 void CComp_TyDeleter::on_delete_comp_type_btn_clicked() {
-  QString sQuery = QString("DELETE FROM components_type "
-                           "WHERE component_type = '%1'")
-                           .arg(m_pUi->comp_type_cBox->currentText());
+  QMessageBox::StandardButton reply;
+  reply = QMessageBox::warning(this, "Warning!", "This action may delete order, are you sure? "
+                                                 "More information is available in the help.",
+                               QMessageBox::Yes | QMessageBox::No);
 
-  try {
-    if (m_hQuery->executeSqlQuery(sQuery)) {
-      CCompsTraits::get_comps().rm(m_pUi->comp_type_cBox->currentText());
+  if (reply == QMessageBox::Yes) {
+    QString sQuery = QString("DELETE FROM components_type "
+                             "WHERE component_type = '%1'")
+                             .arg(m_pUi->comp_type_cBox->currentText());
 
-      QMessageBox::information(this, "Success",
-                               "Component '" + m_pUi->comp_type_cBox->currentText()
-                               + "' was successfully removed!");
+    try {
+      if (m_hQuery->executeSqlQuery(sQuery)) {
+        CCompsTraits::get_comps().rm(m_pUi->comp_type_cBox->currentText());
 
-      sQuery = "SELECT component_type AS 'Component type' FROM components_type;";
+        CMsgBox::show(QMessageBox::Information, this, "Success",
+                      "Component '" + m_pUi->comp_type_cBox->currentText()
+                      + "' was successfully removed!");
 
-      m_hQuery->clear();
-      m_hQuery->executeSqlQuery(sQuery);
+        sQuery = "SELECT component_type AS 'Component type' FROM components_type;";
 
-      QSqlQueryModel* hModel = new QSqlQueryModel();
-      QSortFilterProxyModel* hFilterModel = new QSortFilterProxyModel();
-      hModel->setQuery(m_hQuery->getQuery());
-      hFilterModel->setDynamicSortFilter(true);
-      hFilterModel->setSourceModel(hModel);
-      m_pTbCompType->setModel(hFilterModel);
-      CCompsTraits::get_comps().upd(*m_pUi->comp_type_cBox);
-    } else
-      throw std::invalid_argument("CComp_TyDeleter::on_delete_comp_type_btn_clicked : " +
-                                  m_hQuery->getQuery().lastError().text().toStdString());
-  }
-  catch(const  std::invalid_argument& e) {
-    QMessageBox::critical(this, "Error!", e.what());
-  }
-  catch(...) {
-    QMessageBox::critical(this, "Error!", "CComp_TyDeleter::on_delete_comp_type_btn_clicked : Unexpected error!");
+        m_hQuery->clear();
+        m_hQuery->executeSqlQuery(sQuery);
+
+        QSqlQueryModel* hCompTypeModel = new QSqlQueryModel();
+        QSortFilterProxyModel* hCompTypeFilterModel = new QSortFilterProxyModel();
+
+        hCompTypeModel->setQuery(m_hQuery->getQuery());
+        hCompTypeFilterModel->setDynamicSortFilter(true);
+        hCompTypeFilterModel->setSourceModel(hCompTypeModel);
+        m_pTbCompType->setModel(hCompTypeFilterModel);
+        CCompsTraits::get_comps().upd(*m_pUi->comp_type_cBox);
+
+        sQuery = "SELECT title AS 'Title', component_type AS 'Component type',"
+                 " specifications AS 'Specifications', price AS 'Price', release_date AS 'Release date' "
+                 "FROM components "
+                 "JOIN components_type "
+                 "ON FK_type_code = PK_component_type_id;";
+
+        m_hQuery->clear();
+        m_hQuery->executeSqlQuery(sQuery);
+
+
+        QSqlQueryModel* hCompModel = new QSqlQueryModel();
+        QSortFilterProxyModel* hCompFilterModel = new QSortFilterProxyModel();
+
+        hCompModel->setQuery(m_hQuery->getQuery());
+        hCompFilterModel->setDynamicSortFilter(true);
+        hCompFilterModel->setSourceModel(hCompModel);
+        m_pTbComps->setModel(hCompFilterModel);
+
+        emit updOrders();
+      } else
+        throw std::invalid_argument("CComp_TyDeleter::on_delete_comp_type_btn_clicked : " +
+                                    m_hQuery->getQuery().lastError().text().toStdString());
+    }
+    catch(const  std::invalid_argument& e) {
+      CMsgBox::show(QMessageBox::Critical, this, "Error!", e.what());
+    }
+    catch(...) {
+      CMsgBox::show(QMessageBox::Critical, this, "Error!", "CComp_TyDeleter::on_delete_comp_type_btn_clicked : "
+                                                           "Unexpected error!");
+    }
   }
 }
